Add files via upload
Upload final challenge queries.

I experienced difficulty in running the query for deliverable 2 since my results in the “title” column would differ from the example in the protocol for one or two employees.  I spent over an hour working with a learning assistant (ask-150923) who confirmed that my query and table schemas were correct.  The LA suggested that the example CSV provided in the protocol may be incorrect because the current job title (per the “Title” CSV) for the employee in row 8 is Sr. Engineer, however, the example shows that employee’s job title as Engineer.

I continued to play around with the query and finally received the same result as the example in the protocol by using the employee number from the “Titles” table as my “distinct on” filter.  However, on subsequent runs of that query I noticed that it also returned results that were inconsistent with the example.